requirements.txt
easyocr==1.7.1
ultralytics==8.0.232
torch==1.13.0
torchvision==0.14.0
tensorflow==2.6.0
python-socketio==5.5.0
pytube==15.0.0
pillow==10.2.0
opencv-python==4.6.0.66
onnxruntime==1.16.3
onnx==1.14.0
gdown==4.7.1
Flask-SocketIO==5.1.1
Flask==2.0.1
numpy==1.24.4